refactor(data-view): share animated search control
