+++ apps/web/src/components/data-view/data-view-toolbar.tsx
@@ -1,32 +1,28 @@
 "use client";
 
-import React, { useCallback, useEffect, useRef, useState } from "react";
-import { AnimatePresence, motion } from "framer-motion";
+import React, { useCallback } from "react";
 import { useTranslations } from "next-intl";
-import { Filter, Search, X, RefreshCw } from "lucide-react";
+import { Filter, X, RefreshCw } from "lucide-react";
 import { useQueryClient } from "@tanstack/react-query";
-import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import {
   useDataViewDetail,
   useDataViewList,
   useDataViewSelection,
   useDataViewStatic,
-  useDataViewUrl,
 } from "./use-data-view";
 import { DataViewFilters } from "./data-view-filters";
+import { DataViewSearchControl } from "./data-view-search-control";
 
 // Both toolbar variants use this height so the body area never shifts vertically.
 const TOOLBAR_CLS =
   "flex items-center gap-2 px-3 py-1 border-b bg-background shrink-0 min-h-[3rem]";
-const SEARCH_TRANSITION = { duration: 0.18, ease: [0.22, 1, 0.36, 1] as const };
 
 type DataViewToolbarProps = {
   mode?: "list" | "compact";
 };
 
 export function DataViewToolbar({ mode = "list" }: DataViewToolbarProps) {
-  const { urlState } = useDataViewUrl();
   const { renderToolbarControls, queryKey } = useDataViewStatic();
   const { closeDetail, isClosingDetail } = useDataViewDetail();
   const { listIsTransitioning } = useDataViewList();
@@ -42,91 +38,12 @@ export function DataViewToolbar({ mode = "list" }: DataViewToolbarProps) {
     disableKeepOnlySelected,
     clearSelectedRows,
   } = useDataViewSelection();
-  const [searchOpen, setSearchOpen] = useState(false);
-  const searchInputRef = useRef<HTMLInputElement>(null);
   const t = useTranslations("dataView");
-
-  useEffect(() => {
-    if (mode === "compact") setSearchOpen(false);
-  }, [mode]);
-
-  useEffect(() => {
-    if (searchOpen && searchInputRef.current) {
-      searchInputRef.current.focus();
-    }
-  }, [searchOpen]);
-
-  const handleSearchChange = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      urlState.setSearch(e.target.value);
-    },
-    [urlState]
-  );
-
-  const handleSearchKeyDown = useCallback((e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === "Escape") setSearchOpen(false);
-  }, []);
-
-  const searchControl = (
-    <div className="flex shrink-0 items-center">
-      <AnimatePresence initial={false} mode="popLayout">
-        {searchOpen ? (
-          <motion.div
-            key="search-input"
-            initial={{ width: 36, opacity: 0 }}
-            animate={{ width: 176, opacity: 1 }}
-            exit={{ width: 36, opacity: 0 }}
-            transition={SEARCH_TRANSITION}
-            className="relative shrink-0 overflow-hidden"
-          >
-            <Search className="absolute left-2.5 top-1/2 -translate-y-1/2 h-4 w-4 text-muted-foreground pointer-events-none" />
-            <Input
-              ref={searchInputRef}
-              value={urlState.search}
-              onChange={handleSearchChange}
-              onKeyDown={handleSearchKeyDown}
-              placeholder={t("toolbar.searchPlaceholder")}
-              className="h-9 w-44 pl-8 pr-8 text-sm"
-              aria-label={t("toolbar.searchAria")}
-              data-testid="search-input"
-            />
-            <button
-              onClick={() => setSearchOpen(false)}
-              className="absolute right-2.5 top-1/2 -translate-y-1/2 text-muted-foreground hover:text-foreground"
-              aria-label={t("toolbar.closeSearchAria")}
-            >
-              <X className="h-3.5 w-3.5" />
-            </button>
-          </motion.div>
-        ) : (
-          <motion.div
-            key="search-button"
-            initial={{ width: 32, opacity: 0 }}
-            animate={{ width: 36, opacity: 1 }}
-            exit={{ width: 32, opacity: 0 }}
-            transition={SEARCH_TRANSITION}
-            className="shrink-0 overflow-hidden"
-          >
-            <Button
-              variant="ghost"
-              size="icon"
-              className="h-9 w-9 shrink-0"
-              onClick={() => setSearchOpen(true)}
-              aria-label={t("toolbar.searchAria")}
-              data-testid="search-icon-button"
-            >
-              <Search className="h-4 w-4" />
-            </Button>
-          </motion.div>
-        )}
-      </AnimatePresence>
-    </div>
-  );
 
   if (mode === "list") {
     return (
       <div className={TOOLBAR_CLS} data-testid="toolbar-list">
-        {searchControl}
+        <DataViewSearchControl mode={mode} />
         <DataViewFilters mode="inline" />
         <div className="flex-1" />
         <Button
@@ -177,7 +94,7 @@ export function DataViewToolbar({ mode = "list" }: DataViewToolbarProps) {
 
   return (
     <div className={TOOLBAR_CLS} data-testid="toolbar-detail">
-      {searchControl}
+      <DataViewSearchControl mode={mode} />
       <div className="flex-1" />
       <Button
         variant="ghost"
